# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: ViaSafe API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: reports
    description: Community reports
  - name: alerts
    description: Traffic alerts
  - name: stats
    description: Dashboard statistics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List all reports
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: createReport
      tags: [reports]
      summary: Create a report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReportBody"
      responses:
        "201":
          description: Report created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"

  /reports/{id}:
    patch:
      operationId: updateReport
      tags: [reports]
      summary: Update report status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReportBody"
      responses:
        "200":
          description: Updated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List alerts
      parameters:
        - in: query
          name: active
          schema:
            type: boolean
      responses:
        "200":
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrafficAlert"
    post:
      operationId: createAlert
      tags: [alerts]
      summary: Create an alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlertBody"
      responses:
        "201":
          description: Alert created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrafficAlert"

  /alerts/{id}:
    patch:
      operationId: updateAlert
      tags: [alerts]
      summary: Update alert
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAlertBody"
      responses:
        "200":
          description: Updated alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrafficAlert"

  /stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: Get dashboard summary stats
      responses:
        "200":
          description: Summary statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"

  /stats/timeline:
    get:
      operationId: getStatsTimeline
      tags: [stats]
      summary: Get reports timeline (last 30 days)
      responses:
        "200":
          description: Timeline data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelinePoint"

  /stats/by-type:
    get:
      operationId: getStatsByType
      tags: [stats]
      summary: Get reports grouped by category
      responses:
        "200":
          description: Type breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TypeStat"

  /stats/by-status:
    get:
      operationId: getStatsByStatus
      tags: [stats]
      summary: Get reports grouped by status
      responses:
        "200":
          description: Status breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StatusStat"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    RegisterBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
      required:
        - email
        - password
        - name

    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    UserPublic:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - role
        - createdAt

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserPublic"
      required:
        - token
        - user

    Report:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        category:
          type: string
        description:
          type: string
        location:
          type: string
        lat:
          type: ["number", "null"]
        lon:
          type: ["number", "null"]
        status:
          type: string
        protocol:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - category
        - description
        - location
        - status
        - protocol
        - createdAt
        - updatedAt

    CreateReportBody:
      type: object
      properties:
        userId:
          type: ["integer", "null"]
        category:
          type: string
        description:
          type: string
        location:
          type: string
        lat:
          type: ["number", "null"]
        lon:
          type: ["number", "null"]
        protocol:
          type: string
      required:
        - category
        - description
        - location
        - protocol

    UpdateReportBody:
      type: object
      properties:
        status:
          type: string

    TrafficAlert:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        title:
          type: string
        address:
          type: string
        lat:
          type: number
        lon:
          type: number
        severity:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - type
        - title
        - address
        - lat
        - lon
        - severity
        - active
        - createdAt
        - updatedAt

    CreateAlertBody:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        address:
          type: string
        lat:
          type: number
        lon:
          type: number
        severity:
          type: string
      required:
        - type
        - title
        - address
        - lat
        - lon

    UpdateAlertBody:
      type: object
      properties:
        active:
          type: boolean
        severity:
          type: string

    Stats:
      type: object
      properties:
        totalReports:
          type: integer
        activeAlerts:
          type: integer
        resolvedReports:
          type: integer
        pendingReports:
          type: integer
        deathsIn2026:
          type: integer
        reportsThisWeek:
          type: integer
        reportsToday:
          type: integer
      required:
        - totalReports
        - activeAlerts
        - resolvedReports
        - pendingReports
        - deathsIn2026
        - reportsThisWeek
        - reportsToday

    TimelinePoint:
      type: object
      properties:
        date:
          type: string
        count:
          type: integer
      required:
        - date
        - count

    TypeStat:
      type: object
      properties:
        category:
          type: string
        count:
          type: integer
      required:
        - category
        - count

    StatusStat:
      type: object
      properties:
        status:
          type: string
        count:
          type: integer
      required:
        - status
        - count