# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ViaSafe API specification
 * OpenAPI spec version: 0.2.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register a new user
 */
export const RegisterBody = zod.object({
  email: zod.string(),
  password: zod.string(),
  name: zod.string(),
});

/**
 * @summary Login
 */
export const LoginBody = zod.object({
  email: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  token: zod.string(),
  user: zod.object({
    id: zod.number(),
    email: zod.string(),
    name: zod.string(),
    role: zod.string(),
    createdAt: zod.string(),
  }),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  email: zod.string(),
  name: zod.string(),
  role: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary List all reports
 */
export const ListReportsQueryParams = zod.object({
  status: zod.coerce.string().optional(),
  category: zod.coerce.string().optional(),
  limit: zod.coerce.number().optional(),
});

export const ListReportsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number().nullish(),
  category: zod.string(),
  description: zod.string(),
  location: zod.string(),
  lat: zod.number().nullish(),
  lon: zod.number().nullish(),
  status: zod.string(),
  protocol: zod.string(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListReportsResponse = zod.array(ListReportsResponseItem);

/**
 * @summary Create a report
 */
export const CreateReportBody = zod.object({
  userId: zod.number().nullish(),
  category: zod.string(),
  description: zod.string(),
  location: zod.string(),
  lat: zod.number().nullish(),
  lon: zod.number().nullish(),
  protocol: zod.string(),
});

/**
 * @summary Update report status
 */
export const UpdateReportParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateReportBody = zod.object({
  status: zod.string().optional(),
});

export const UpdateReportResponse = zod.object({
  id: zod.number(),
  userId: zod.number().nullish(),
  category: zod.string(),
  description: zod.string(),
  location: zod.string(),
  lat: zod.number().nullish(),
  lon: zod.number().nullish(),
  status: zod.string(),
  protocol: zod.string(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary List alerts
 */
export const ListAlertsQueryParams = zod.object({
  active: zod.coerce.boolean().optional(),
});

export const ListAlertsResponseItem = zod.object({
  id: zod.number(),
  type: zod.string(),
  title: zod.string(),
  address: zod.string(),
  lat: zod.number(),
  lon: zod.number(),
  severity: zod.string(),
  active: zod.boolean(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListAlertsResponse = zod.array(ListAlertsResponseItem);

/**
 * @summary Create an alert
 */
export const CreateAlertBody = zod.object({
  type: zod.string(),
  title: zod.string(),
  address: zod.string(),
  lat: zod.number(),
  lon: zod.number(),
  severity: zod.string().optional(),
});

/**
 * @summary Update alert
 */
export const UpdateAlertParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateAlertBody = zod.object({
  active: zod.boolean().optional(),
  severity: zod.string().optional(),
});

export const UpdateAlertResponse = zod.object({
  id: zod.number(),
  type: zod.string(),
  title: zod.string(),
  address: zod.string(),
  lat: zod.number(),
  lon: zod.number(),
  severity: zod.string(),
  active: zod.boolean(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Get dashboard summary stats
 */
export const GetStatsResponse = zod.object({
  totalReports: zod.number(),
  activeAlerts: zod.number(),
  resolvedReports: zod.number(),
  pendingReports: zod.number(),
  deathsIn2026: zod.number(),
  reportsThisWeek: zod.number(),
  reportsToday: zod.number(),
});

/**
 * @summary Get reports timeline (last 30 days)
 */
export const GetStatsTimelineResponseItem = zod.object({
  date: zod.string(),
  count: zod.number(),
});
export const GetStatsTimelineResponse = zod.array(GetStatsTimelineResponseItem);

/**
 * @summary Get reports grouped by category
 */
export const GetStatsByTypeResponseItem = zod.object({
  category: zod.string(),
  count: zod.number(),
});
export const GetStatsByTypeResponse = zod.array(GetStatsByTypeResponseItem);

/**
 * @summary Get reports grouped by status
 */
export const GetStatsByStatusResponseItem = zod.object({
  status: zod.string(),
  count: zod.number(),
});
export const GetStatsByStatusResponse = zod.array(GetStatsByStatusResponseItem);